Create Scraping_Numbers_From_HTML
This program uses urllib to read the HTML from the data files, and parse the data, extracting numbers and compute the sum of the numbers in the file. BeautifulSoup was also used, which is a python library that removes HTML markup and saves the content for easier parsing of information.