Create Following_html_links
The program uses urllib to read the HTML from the data file, extract the href= vaues from the anchor tags, scan for a tag that is in a particular position relative to the first name in the list, follows that link and repeats the process a number of times and report the last name found

Assignment:
Find the link at position 18 (the first name is 1). Follow that link. Repeat this process 7 times. The answer is the last name that you retrieve.